pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["Author <author@email.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "^2.1.2"
numpy = "^1.26.1"
matplotlib = "^3.8.2"
pyyaml = "^6.0.1"
snakemake = "^7.32.4"
pycubexr = "^1.2.1"
pygments = "^2.16.1"
pulp = "2.7.0"
bs4 = "^0.0.2"
lxml = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
